requirements.txt
# ========================================================
# DEPENDÊNCIAS DO PROJETO POSTECH-CHALLENGE
# ========================================================
# Para instalar todas as dependências de produção:
# pip install -r requirements.txt
#
# Para desenvolvimento (incluindo dependências opcionais):
# pip install -r requirements-dev.txt
# ========================================================

# ==================== DEPENDÊNCIAS CORE ====================
# Essenciais para o funcionamento básico do projeto
pandas>=2.0.0            # Manipulação avançada de dados
numpy>=1.24.0            # Computação numérica eficiente
python-dotenv            # Gerenciamento de variáveis de ambiente

# ================== COLETA E PROCESSAMENTO =================
yfinance                 # API para dados de mercado
lxml                     # Parser HTML/XML para web scraping
pandas-ta                # Indicadores técnicos para análise

# ===================== BANCO DE DADOS ======================
duckdb                   # Banco de dados OLAP embutido
duckdb-engine            # Interface SQLAlchemy para DuckDB

# ==================== MACHINE LEARNING =====================
scikit-learn             # Algoritmos ML tradicionais
lightgbm                 # Gradient Boosting eficiente
tensorflow               # Redes neurais e modelos sequenciais
optuna                   # Otimização de hiperparâmetros

# ====================== MONITORAMENTO ======================
mlflow                   # Rastreamento de experimentos ML

# ======================== API/SERVIÇO ======================
fastapi                  # Framework para construção de APIs
uvicorn[standard]        # Servidor ASGI de alta performance

# ====================== DEPENDÊNCIAS DEV ===================
# (Movidas para requirements-dev.txt - ver abaixo)